src/test/java/io/github/vlad324/n312/SolutionTest.java
package io.github.vlad324.n312;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {

    private final Solution solution = new Solution();

    private static Stream<Arguments> testDataProvider() {
        return Stream.<Arguments>builder()
            .add(arguments(new int[]{3, 1, 5, 8}, 167))
            .add(arguments(new int[]{1, 5}, 10))
            .add(arguments(new int[]{1}, 1))
            .build();
    }

    @ParameterizedTest
    @MethodSource("testDataProvider")
    void should_produce_expected_result(int[] nums, int expected) {
        // when
        final var actual = solution.maxCoins(nums);

        // then
        assertThat(actual).isEqualTo(expected);
    }

}